Meta folder loc fix
