+++ src/meta/army_list.ts
@@ -7,11 +7,11 @@ import {
   IRONJAWZ,
   KHARADRON_OVERLORDS,
   KHORNE,
-  LEGIONS_OF_NAGASH,
   NIGHTHAUNT,
   SERAPHON,
   SKAVEN,
   SYLVANETH,
+  STORMCAST_ETERNALS,
   TZEENTCH,
   TSupportedFaction,
 } from './factions'
@@ -26,10 +26,10 @@ import IdonethDeepkin from 'army/idoneth'
 import Ironjawz from 'army/ironjawz'
 import KharadronOverlords from 'army/kharadron_overlords'
 import Khorne from 'army/khorne'
-import LegionsofNagash from 'army/legions_of_nagash'
 import Nighthaunt from 'army/nighthaunt'
 import Seraphon from 'army/seraphon'
 import Skaven from 'army/skaven'
+import StormcastEternals from 'army/stormcast_eternals'
 import Sylvaneth from 'army/sylvaneth'
 import Tzeentch from 'army/tzeentch'
 
@@ -66,10 +66,6 @@ const ArmyList: TArmyList = {
     Army: { ...Khorne },
     GrandAlliance: CHAOS,
   },
-  [LEGIONS_OF_NAGASH]: {
-    Army: { ...LegionsofNagash },
-    GrandAlliance: DEATH,
-  },
   [NIGHTHAUNT]: {
     Army: { ...Nighthaunt },
     GrandAlliance: DEATH,
@@ -84,6 +80,10 @@ const ArmyList: TArmyList = {
   },
   [SYLVANETH]: {
     Army: { ...Sylvaneth },
+    GrandAlliance: ORDER,
+  },
+  [STORMCAST_ETERNALS]: {
+    Army: { ...StormcastEternals },
     GrandAlliance: ORDER,
   },
   [TZEENTCH]: {
